Remove `/settings timezone`, and move `/notifications` and `/roles` under new `/setup-feature` group cog.
