+++ cogs/setup_features.py
@@ -5,14 +5,16 @@ from discord import app_commands
 from discord.ext import commands
 
 from middleware import defer_interaction, ensure_server_document
+from ui.embeds.roles import roles_menu_embed
 from ui.views.notifications import SelectOperatorView
+from ui.views.roles import RolesView
 
 if TYPE_CHECKING:
     # To prevent circular imports
     from bot import DiscordBot
 
 
-class NotificationsCog(commands.Cog):
+class SetupFeatureGroupCog(commands.GroupCog, name="setup-feature"):
     def __init__(self, bot: "DiscordBot") -> None:
         self.bot = bot
 
@@ -42,6 +44,17 @@ class NotificationsCog(commands.Cog):
             view=SelectOperatorView(self.bot, channel),
         )
 
+    @app_commands.command(name="roles")
+    @commands.has_permissions(administrator=True)
+    @app_commands.checks.bot_has_permissions(manage_roles=True)
+    @defer_interaction(ephemeral_default=True)
+    @ensure_server_document
+    async def roles(self, interaction: discord.Interaction) -> None:
+        """
+        Admins only: Enable or disable CodeGrind roles
+        """
+        await interaction.followup.send(embed=roles_menu_embed(), view=RolesView())
+
 
 async def setup(bot: "DiscordBot") -> None:
-    await bot.add_cog(NotificationsCog(bot))
+    await bot.add_cog(SetupFeatureGroupCog(bot))
